pyproject.toml
[tool.poetry]
name = "urobor"
version = "0.1.0"
description = ""
authors = ["Dawid Kraczkowski <dawid.kraczkowski@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
marko = "^1.2.1"
bson = "^0.5.10"
mistletoe = "^0.8.2"
mistune = "^2.0.2"
markdown-it-py = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
